Added wilks and partial input
Partial input is save in case it contains a '.' such as "67."
This input would be converted to a float correctly but will
strip the "." from it and make it hard to enter such numbers
in the field since the content of the field was controlled by
the model. With the partial model we can now save a partial input
that can eventually lead to a correct input.